+++ examples/lib/commons/square_component.dart
@@ -4,11 +4,9 @@ import 'package:flame/components.dart';
 import 'package:flame/palette.dart';
 
 class SquareComponent extends PositionComponent {
-  Paint paint = BasicPalette.white.paint;
+  Paint paint = BasicPalette.white.paint();
 
-  SquareComponent() {
-    size = Vector2.all(100.0);
-  }
+  SquareComponent() : super(size: Vector2.all(100.0));
 
   @override
   void render(Canvas c) {