+++ examples/lib/stories/controls/overlapping_tapables.dart
@@ -14,18 +14,17 @@ class TapableSquare extends PositionComponent with Tapable {
       rng.nextInt(256),
       0.9,
     );
-    return PaletteEntry(color).paint;
+    return PaletteEntry(color).paint();
   }
 
   Paint currentPaint;
 
-  TapableSquare({Vector2 position})
-      : super(
+  TapableSquare({Vector2? position})
+      : currentPaint = _randomPaint(),
+        super(
           position: position ?? Vector2.all(100),
           size: Vector2.all(100),
-        ) {
-    currentPaint = _randomPaint();
-  }
+        );
 
   @override
   void render(Canvas canvas) {