+++ examples/lib/stories/controls/scroll.dart
@@ -5,19 +5,17 @@ import 'package:flame/gestures.dart';
 import 'package:flame/palette.dart';
 import 'package:flame/extensions.dart';
 
-// TODO(luan) doesnt seem to be working, figure it out
 class ScrollGame extends BaseGame with ScrollDetector {
   static const speed = 2000.0;
   final _size = Vector2.all(50);
-  final _paint = BasicPalette.white.paint;
+  final _paint = BasicPalette.white.paint();
 
   Vector2 position = Vector2.all(100);
-  Vector2 target;
+  Vector2? target;
 
   @override
   void onScroll(PointerScrollEvent event) {
-    print(event.scrollDelta);
-    target = position - event.scrollDelta.toVector2() * 10;
+    target = position + event.scrollDelta.toVector2() * 5;
   }
 
   @override
@@ -29,9 +27,18 @@ class ScrollGame extends BaseGame with ScrollDetector {
   @override
   void update(double dt) {
     super.update(dt);
+    final target = this.target;
+    final ds = speed * dt;
     if (target != null) {
-      final dir = (target - position).normalized();
-      position += dir * (speed * dt);
+      if (position != target) {
+        final diff = target - position;
+        if (diff.length < ds) {
+          position.setFrom(target);
+        } else {
+          diff.scaleTo(ds);
+          position.setFrom(position + diff);
+        }
+      }
     }
   }
 }