+++ examples/lib/stories/effects/combined_effect.dart
@@ -10,17 +10,17 @@ final green = Paint()..color = const Color(0xAA338833);
 final red = Paint()..color = const Color(0xAA883333);
 
 class CombinedEffectGame extends BaseGame with TapDetector {
-  SquareComponent greenSquare, redSquare;
+  late SquareComponent greenSquare, redSquare;
 
   @override
   Future<void> onLoad() async {
     greenSquare = SquareComponent()
       ..paint = green
-      ..position = Vector2.all(100);
+      ..position.setValues(100, 100);
 
     redSquare = SquareComponent()
       ..paint = red
-      ..position = Vector2.all(100);
+      ..position.setValues(100, 100);
 
     add(greenSquare);
     add(redSquare);