+++ examples/lib/stories/effects/infinite_effect.dart
@@ -15,13 +15,13 @@ final orange = Paint()..color = const Color(0xAABB6633);
 SquareComponent makeSquare(Paint paint) {
   return SquareComponent()
     ..paint = paint
-    ..position = Vector2.all(100);
+    ..position.setValues(100, 100);
 }
 
 class InfiniteEffectGame extends BaseGame with TapDetector {
-  SquareComponent greenSquare;
-  SquareComponent redSquare;
-  SquareComponent orangeSquare;
+  late SquareComponent greenSquare;
+  late SquareComponent redSquare;
+  late SquareComponent orangeSquare;
 
   @override
   Future<void> onLoad() async {