+++ examples/lib/stories/effects/move_effect.dart
@@ -7,13 +7,12 @@ import 'package:flutter/material.dart';
 import '../../commons/square_component.dart';
 
 class MoveEffectGame extends BaseGame with TapDetector {
-  SquareComponent square;
+  late SquareComponent square;
 
   @override
   Future<void> onLoad() async {
-    add(
-      square = SquareComponent()..position = Vector2.all(100),
-    );
+    square = SquareComponent()..position.setValues(100, 100);
+    add(square);
   }
 
   @override