+++ packages/flame/lib/src/game/camera.dart
@@ -1,7 +1,8 @@
 import 'dart:math' as math;
-import 'dart:ui' show Rect;
+import 'dart:ui' show Rect, Canvas;
 
 import '../../components.dart';
+import '../../extensions.dart';
 import '../../game.dart';
 
 /// Utility method to smoothly transition vectors.
@@ -35,7 +36,7 @@ void _moveToTarget(
 /// There are three major factors that determine the camera position:
 ///
 /// * Follow
-/// If you want, you can call [followObject] at the beginning of your
+/// If you want, you can call [followComponent] at the beginning of your
 /// stage/world/level, and provided a [PositionComponent].
 /// The camera will follow this component making sure its position is fixed
 /// on the screen.
@@ -89,29 +90,36 @@ class Camera {
   double cameraSpeed = defaultCameraSpeed;
   double shakeIntensity = defaultShakeIntensity;
 
-  /// This is the current position of the camera, ie the world coordinate that is
-  /// rendered on the top left of the screen (origin of the screen space).
+  /// This is the current position of the camera, ie the world coordinate that
+  /// is rendered on the top left of the screen (origin of the screen space).
   ///
   /// Zero means no translation is applied.
   /// You can't change this directly; the camera will handle all ongoing
   /// movements so they smoothly transition.
   /// If you want to immediately snap the camera to a new place, you can do:
   /// ```
-  ///   camera.moveTo(newPosition);
-  ///   camera.snap();
+  ///   camera.snapTo(newPosition);
   /// ```
-  Vector2 get position => _position.clone();
+  Vector2 get position => _internalPosition.clone();
 
-  final Vector2 _position = Vector2.zero();
+  /// Do not change this directly since it bypasses [onPositionUpdate]
+  final Vector2 _internalPosition = Vector2.zero();
 
-  /// If set, the camera will "follow" this component, making sure that this
-  /// component is always rendered in a fixed position in the screen, by
-  /// immediately moving the camera to "focus" on the object.
+  Vector2 get _position => _internalPosition;
+  set _position(Vector2 position) {
+    _internalPosition.setFrom(position);
+    onPositionUpdate(_internalPosition);
+  }
+
+  /// If set, the camera will "follow" this vector, making sure that this
+  /// vector is always rendered in a fixed position in the screen, by
+  /// immediately moving the camera to "focus" on the where the vector is.
   ///
-  /// You probably want to set it to the player component.
-  /// Note that this is not smooth because the movement of the follow object
+  /// You might want to set it to the player component by using the
+  /// [followComponent] method.
+  /// Note that this is not smooth because the movement of the followed vector
   /// is assumed to be smooth.
-  PositionComponent? follow;
+  Vector2? follow;
 
   /// Where in the screen the follow object should be.
   ///
@@ -135,7 +143,32 @@ class Camera {
   /// add any non-smooth movement.
   Rect? worldBounds;
 
+  /// If set, the camera will zoom by this ratio. This can be greater than 1 (zoom in)
+  /// or smaller (zoom out), but should always be greater than zero.
+  ///
+  /// Note: do not confuse this with the zoom applied by the viewport. The
+  /// viewport applies a (normally) fixed zoom to adapt multiple screens into
+  /// one aspect ratio. The zoom might be different per dimension depending
+  /// on the Viewport implementation. Also, if used with the default
+  /// BaseGame implementation, it will apply to all components.
+  /// The zoom from the camera is only for components that respect camera,
+  /// and is applied after the viewport is set. It exists to be used if there
+  /// is any kind of user configurable camera on your game.
+  double zoom = 1.0;
+
   Camera();
+
+  /// Use this method to transform the canvas using the current rules provided
+  /// by this camera object.
+  ///
+  /// If you are using BaseGame, this will be done for you for all non-HUD
+  /// components.
+  /// When using this method you are responsible for saving/restoring canvas
+  /// state to avoid leakage.
+  void apply(Canvas canvas) {
+    canvas.translateVector(-position);
+    canvas.scale(zoom);
+  }
 
   /// This smoothly updates the camera for an amount of time [dt].
   ///
@@ -144,13 +177,11 @@ class Camera {
     final ds = cameraSpeed * dt;
     final shake = Vector2(_shakeDelta(), _shakeDelta());
 
+    _moveToTarget(_currentRelativeOffset, _targetRelativeOffset, ds);
     if (_targetCameraDelta != null && _currentCameraDelta != null) {
       _moveToTarget(_currentCameraDelta!, _targetCameraDelta!, ds);
-      _position.setFrom(_currentCameraDelta! + shake);
-    } else {
-      _moveToTarget(_currentRelativeOffset, _targetRelativeOffset, ds);
-      _position.setFrom(_getTarget() + shake);
     }
+    _position = _target() + shake;
 
     if (shaking) {
       _shakeTimer -= dt;
@@ -168,37 +199,58 @@ class Camera {
       _currentCameraDelta!.setFrom(_targetCameraDelta!);
     }
     _currentRelativeOffset.setFrom(_targetRelativeOffset);
+    update(0);
   }
 
   /// Converts a vector in the screen space to the world space.
   Vector2 screenToWorld(Vector2 screenCoordinates) {
-    return screenCoordinates + _position;
+    return (screenCoordinates + _position) / zoom;
   }
 
   /// Converts a vector in the world space to the screen space.
   Vector2 worldToScreen(Vector2 worldCoordinates) {
-    return worldCoordinates - _position;
+    return worldCoordinates * zoom - _position;
   }
 
   // Follow
 
-  /// Immediately snaps the camera to start following the object [follow].
+  /// Immediately snaps the camera to start following the [component].
   ///
-  /// This means that the camera will move so that the [follow] object is
-  /// in a fixed position on the screen.
+  /// This means that the camera will move so that the position vector of the
+  /// component is in a fixed position on the screen.
   /// That position is determined by a fraction of screen size defined by
   /// [relativeOffset] (default to the center).
   /// [worldBounds] can be optionally set to add boundaries to how far the
   /// camera is allowed to move.
-  /// The object is "grabbed" by its anchor (default top left). So for example
-  /// if you want the center of the object to be at the fixed position, set
-  /// its anchor to center.
-  void followObject(
-    PositionComponent follow, {
+  /// The component is "grabbed" by its anchor (default top left).
+  /// So for example if you want the center of the object to be at the fixed
+  /// position, set the components anchor to center.
+  void followComponent(
+    PositionComponent component, {
     Vector2? relativeOffset,
     Rect? worldBounds,
   }) {
-    this.follow = follow;
+    followVector2(
+      component.position,
+      relativeOffset: relativeOffset,
+      worldBounds: worldBounds,
+    );
+  }
+
+  /// Immediately snaps the camera to start following [vector2].
+  ///
+  /// This means that the camera will move so that the position vector is in a
+  /// fixed position on the screen.
+  /// That position is determined by a fraction of screen size defined by
+  /// [relativeOffset] (default to the center).
+  /// [worldBounds] can be optionally set to add boundaries to how far the
+  /// camera is allowed to move.
+  void followVector2(
+    Vector2 vector2, {
+    Vector2? relativeOffset,
+    Rect? worldBounds,
+  }) {
+    follow = vector2;
     this.worldBounds = worldBounds;
     _targetRelativeOffset.setFrom(relativeOffset ?? Anchor.center.toVector2());
     _currentRelativeOffset.setFrom(_targetRelativeOffset);
@@ -214,12 +266,13 @@ class Camera {
     _targetRelativeOffset.setFrom(newRelativeOffset);
   }
 
-  Vector2 _getTarget() {
-    if (follow == null) {
-      return Vector2.zero();
-    }
-    final screenDelta = gameRef.size.clone()..multiply(_currentRelativeOffset);
-    final attemptedTarget = follow!.position - screenDelta;
+  Vector2 _screenDelta() {
+    return gameRef.size.clone()..multiply(_currentRelativeOffset);
+  }
+
+  Vector2 _target() {
+    final target = _currentCameraDelta ?? follow ?? Vector2.zero();
+    final attemptedTarget = target - _screenDelta();
 
     final bounds = worldBounds;
     if (bounds != null) {
@@ -259,9 +312,16 @@ class Camera {
   ///
   /// The camera will be smoothly transitioned to this position.
   /// This will replace any previous targets.
-  void moveTo(Vector2 p) {
-    _currentCameraDelta = _position;
-    _targetCameraDelta = p.clone();
+  void moveTo(Vector2 position) {
+    _currentCameraDelta = _position + _screenDelta();
+    _targetCameraDelta = position.clone();
+  }
+
+  /// Instantly moves the camera to the target, bypassing follow.
+  /// This will replace any previous targets.
+  void snapTo(Vector2 position) {
+    moveTo(position);
+    snap();
   }
 
   /// Smoothly resets any moveTo targets.
@@ -290,4 +350,8 @@ class Camera {
     }
     return 0.0;
   }
+
+  /// If you need updated on when the position of the camera is updated you
+  /// can override this.
+  void onPositionUpdate(Vector2 position) {}
 }